backend/requirements.txt
Django==5.0.1
django-ninja==1.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1
